added llama_cpp:builder

+++ packages/llm/llama_cpp/install.sh
@@ -1,11 +1,6 @@
 #!/usr/bin/env bash
 set -ex
 
-if [ "$FORCE_BUILD" == "on" ]; then
-	echo "Forcing build of llama.cpp ${LLAMA_CPP_VERSION}"
-	exit 1
-fi
-
 pip3 install --no-cache-dir --verbose \
         typing-extensions \
         uvicorn \
@@ -16,4 +11,9 @@ pip3 install --no-cache-dir --verbose \
         fastapi \
         pydantic-settings
         
+if [ "$FORCE_BUILD" == "on" ]; then
+	echo "Forcing build of llama.cpp ${LLAMA_CPP_VERSION}"
+	exit 1
+fi
+   
 pip3 install --no-cache-dir --verbose llama-cpp-python==${LLAMA_CPP_VERSION}
